+++ fw/src/utils_function.h
@@ -21,7 +21,7 @@ uint32_t convert_string_to_hex(char src_string[]);
 uint64_t convert_string_to_hex_uint64_t(char src_string[]);
 uint32_t system_call(char cmd[]); //todo create a generic function to create a system call with a pipe.
 
-uint32_t encode_uint_data_to_send(uint32_t data_as_uint);
+float encode_uint_data_to_send(float data_as_uint);
 void build_data_packet_header(char* header, uint32_t len, int type);
 
 float convert_fix_point_to_float(uint32_t fixed_point);
